ci(tangle-dapp): Fix build errors

+++ libs/tangle-shared-ui/src/queries/restake/rewardVault.ts
@@ -1,10 +1,10 @@
 import type { ApiRx } from '@polkadot/api';
-import type { Option, StorageKey, u128, Vec } from '@polkadot/types';
+import type { Option, StorageKey, u32, Vec } from '@polkadot/types';
 import { TanglePrimitivesServicesAsset } from '@polkadot/types/lookup';
 import { map, of } from 'rxjs';
 
 function toPrimitive(
-  entries: [StorageKey<[u128]>, Option<Vec<TanglePrimitivesServicesAsset>>][],
+  entries: [StorageKey<[u32]>, Option<Vec<TanglePrimitivesServicesAsset>>][],
 ): [vaultId: bigint, assetIds: bigint[] | null][] {
   return entries.map(([vaultId, assets]) => {
     const vaultIdBigInt = vaultId.args[0].toBigInt();
